Delete unused components, fix lint errors
