+++ typescript/packages/jumble/src/contexts/CharmManagerContext.tsx
@@ -14,19 +14,19 @@ const CharmManagerContext = createContext<CharmManagerContextType>({
 
 export const CharmsManagerProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const { replicaName } = useParams<{ replicaName: string }>();
+  const [effectiveReplica, setEffectiveReplica] = React.useState<string>(
+    () => localStorage.getItem("lastReplica") || "common-knowledge"
+  );
 
-  console.log("CharmManagerProvider", replicaName);
+  React.useEffect(() => {
+    console.log("CharmManagerProvider", replicaName);
 
-  let effectiveReplica: string;
-  if (replicaName) {
-    // When a replica is provided in the URL, use it and save it as the last visited
-    effectiveReplica = replicaName;
-    localStorage.setItem("lastReplica", replicaName);
-  } else {
-    // Otherwise, pull the last visited replica from local storage.
-    // Falling back to "common-knowledge" if nothing was stored.
-    effectiveReplica = localStorage.getItem("lastReplica") || "common-knowledge";
-  }
+    if (replicaName) {
+      // When a replica is provided in the URL, use it and save it as the last visited
+      setEffectiveReplica(replicaName);
+      localStorage.setItem("lastReplica", replicaName);
+    }
+  }, [replicaName]);
 
   const charmManager = useMemo(() => {
     return new CharmManager(effectiveReplica);
