+++ typescript/packages/jumble/src/contexts/LanguageModelContext.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useState } from "react";
+import React, { createContext, useCallback, useContext, useState } from "react";
 
 interface LanguageModelContextType {
   modelId: string | null;
@@ -7,15 +7,15 @@ interface LanguageModelContextType {
 
 const LanguageModelContext = createContext<LanguageModelContextType>({
   modelId: "groq:llama-3.3-70b-versatile",
-  setPreferredModel: () => {},
+  setPreferredModel: () => { },
 });
 
 export const LanguageModelProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const [modelId, setModelId] = useState<string | null>("groq:llama-3.3-70b-versatile");
 
-  const setPreferredModel = (modelId: string) => {
+  const setPreferredModel = useCallback((modelId: string) => {
     setModelId(modelId);
-  };
+  }, []);
 
   return (
     <LanguageModelContext.Provider value={{ modelId, setPreferredModel }}>
